♻️ move ProjectList into dedicated component
