+++ frontend/src/pages/Projects/AllProjects/index.tsx
@@ -1,0 +1,75 @@
+import { gql } from "@apollo/client";
+import { sortBy } from "lodash";
+import { useMemo } from "react";
+import { generatePath, Link } from "react-router-dom";
+import { RoutePaths } from "src/App";
+import ProjectCard from "src/components/ProjectCard";
+import QueryWrapper from "src/components/QueryWrapper";
+import { GITHUB_REPO_FIELDS_FOR_PROJECT_CARD_FRAGMENT } from "src/graphql/fragments";
+import { useAuth } from "src/hooks/useAuth";
+import { useHasuraQuery } from "src/hooks/useHasuraQuery";
+import { HasuraUserRole } from "src/types";
+import { GetProjectsQuery } from "src/__generated/graphql";
+import { Project } from "..";
+
+export default function AllProjects() {
+  const { ledProjectIds, githubUserId } = useAuth();
+
+  const getProjectsQuery = useHasuraQuery<GetProjectsQuery>(GET_PROJECTS_QUERY, HasuraUserRole.Public, {
+    variables: { githubUserId },
+  });
+
+  const projects = useMemo(
+    () => sortBy(getProjectsQuery.data?.projects, p => !p.pendingInvitations.length),
+    [getProjectsQuery.data?.projects]
+  );
+
+  const isProjectMine = (project: Project) =>
+    ledProjectIds.includes(project?.id) || project?.pendingInvitations?.length > 0;
+
+  return (
+    <QueryWrapper query={getProjectsQuery}>
+      <div className="flex flex-col gap-5">
+        {projects &&
+          projects.map(project => (
+            <Link
+              key={project.id}
+              to={generatePath(isProjectMine(project) ? RoutePaths.MyProjectDetails : RoutePaths.ProjectDetails, {
+                projectId: project.id,
+              })}
+            >
+              <ProjectCard {...project} />
+            </Link>
+          ))}
+      </div>
+    </QueryWrapper>
+  );
+}
+
+export const GET_PROJECTS_QUERY = gql`
+  ${GITHUB_REPO_FIELDS_FOR_PROJECT_CARD_FRAGMENT}
+  query GetProjects($githubUserId: bigint = 0) {
+    projects {
+      id
+      name
+      totalSpentAmountInUsd
+      projectDetails {
+        description
+        telegramLink
+        logoUrl
+      }
+      pendingInvitations(where: { githubUserId: { _eq: $githubUserId } }) {
+        id
+      }
+      projectLeads {
+        user {
+          displayName
+          avatarUrl
+        }
+      }
+      githubRepo {
+        ...GithubRepoFieldsForProjectCard
+      }
+    }
+  }
+`;
